requirements.txt
PyMuPDF==1.18.19
google-api-core==2.3.0
google-generativeai==0.1.0